Rename vehicle package to vehicles to be consistant with packages (which can't simply be named package, because that's a keyword).
