+++ external-resolver/src/main/scala/org/genivi/sota/resolver/FilterDirectives.scala
@@ -16,7 +16,7 @@ import org.genivi.sota.marshalling.CirceMarshallingSupport._
 import org.genivi.sota.resolver.db.{Filters, PackageFilters}
 import org.genivi.sota.resolver.types.{Filter, PackageFilter}
 import org.genivi.sota.resolver.packages._
-import org.genivi.sota.resolver.vehicle._
+import org.genivi.sota.resolver.vehicles._
 import org.genivi.sota.rest.Validation._
 import scala.concurrent.ExecutionContext
 import scala.concurrent.Future
