+++ external-resolver/src/main/scala/org/genivi/sota/resolver/vehicles/VehicleDirectives.scala
@@ -2,7 +2,7 @@
  * Copyright: Copyright (C) 2015, Jaguar Land Rover
  * License: MPL-2.0
  */
-package org.genivi.sota.resolver.vehicle
+package org.genivi.sota.resolver.vehicles
 
 import akka.http.scaladsl.model.StatusCodes
 import akka.http.scaladsl.model.StatusCodes.NoContent
