+++ external-resolver/src/main/scala/org/genivi/sota/resolver/vehicles/VehicleFunctions.scala
@@ -2,7 +2,7 @@
  * Copyright: Copyright (C) 2015, Jaguar Land Rover
  * License: MPL-2.0
  */
-package org.genivi.sota.resolver.vehicle
+package org.genivi.sota.resolver.vehicles
 
 import org.genivi.sota.resolver.packages.{Package, PackageFunctions}
 import org.genivi.sota.resolver.Errors
