+++ external-resolver/src/test/scala/org/genivi/sota/resolver/test/Requests.scala
@@ -14,7 +14,7 @@ import org.genivi.sota.marshalling.CirceMarshallingSupport
 import CirceMarshallingSupport._
 import org.genivi.sota.resolver.DependenciesDirectives.makeFakeDependencyMap
 import org.genivi.sota.resolver.packages.Package
-import org.genivi.sota.resolver.vehicle.Vehicle
+import org.genivi.sota.resolver.vehicles.Vehicle
 import org.genivi.sota.resolver.types.{Filter, PackageFilter}
 import org.scalatest.Matchers
 import scala.concurrent.duration._
