+++ external-resolver/src/test/scala/org/genivi/sota/resolver/test/ResolveResourceSpec.scala
@@ -5,7 +5,7 @@ import eu.timepit.refined.Refined
 import io.circe.generic.auto._
 import org.genivi.sota.marshalling.CirceMarshallingSupport._
 import org.genivi.sota.resolver.packages.Package
-import org.genivi.sota.resolver.vehicle.Vehicle
+import org.genivi.sota.resolver.vehicles.Vehicle
 import org.genivi.sota.resolver.types.PackageFilter
 import org.genivi.sota.resolver.Errors.Codes
 import org.genivi.sota.rest.{ErrorRepresentation, ErrorCodes}
