+++ external-resolver/src/test/scala/org/genivi/sota/resolver/test/VehiclesResourceSpec.scala
@@ -12,7 +12,7 @@ import org.genivi.sota.marshalling.CirceMarshallingSupport
 import CirceMarshallingSupport._
 import org.genivi.sota.resolver.Errors.Codes
 import org.genivi.sota.resolver.packages.Package
-import org.genivi.sota.resolver.vehicle.Vehicle
+import org.genivi.sota.resolver.vehicles.Vehicle
 import org.genivi.sota.resolver.types.PackageFilter
 import org.genivi.sota.rest.{ErrorCodes, ErrorRepresentation}
 import org.scalacheck._
